Try with independent python file to get first env
