+++ environment.yml
@@ -2,10 +2,10 @@ name: geoutils
 channels:
   - conda-forge
 dependencies:
-  - python >= 3.8
-  - geopandas >= 0.10.0
+  - python>=3.8
+  - geopandas>=0.10.0
   - matplotlib
   - pyproj
-  - rasterio >= 1.3
+  - rasterio>=1.3
   - scipy
   - tqdm
